SH_issue_111 Fix #111 Update and move MailConfig.java to tools package. Delete MyConstants.java
