--- skillful_network_server/src/main/java/fr/uca/cdr/skillful_network/configurations/MyConstants.java
@@ -1,12 +1,0 @@
-package fr.uca.cdr.skillful_network.configurations;
-public class MyConstants {
-	 
-    // Replace with your email here:  
-    public static final String MY_EMAIL = "poec2020.softeam@gmail.com";
- 
-    // Replace password!!
-    public static final String MY_PASSWORD = "123456softeam";
- 
-
- 
-}
