+++ skillful_network_server/src/main/java/fr/uca/cdr/skillful_network/tools/MailConfig.java
@@ -1,0 +1,36 @@
+package fr.uca.cdr.skillful_network.tools;
+
+import java.util.Properties;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.mail.javamail.JavaMailSender;
+import org.springframework.mail.javamail.JavaMailSenderImpl;
+
+@Configuration
+public class MailConfig {
+	// Replace with your email here:
+	public static final String MY_EMAIL = "poec2020.softeam@gmail.com";
+
+	// Replace password!!
+	public static final String MY_PASSWORD = "123456softeam";
+
+	@Bean
+	public JavaMailSender getJavaMailSender() {
+		JavaMailSenderImpl mailSender = new JavaMailSenderImpl();
+		mailSender.setHost("smtp.gmail.com");
+		mailSender.setPort(587);
+
+		mailSender.setUsername(MY_EMAIL);
+		mailSender.setPassword(MY_PASSWORD);
+
+		Properties props = mailSender.getJavaMailProperties();
+		props.put("mail.transport.protocol", "smtp");
+		props.put("mail.smtp.auth", "true");
+		props.put("mail.smtp.starttls.enable", "true");
+		props.put("mail.debug", "true");
+
+		return mailSender;
+	}
+
+}
